Refactor counter imports
