+++ packages/client/src/modules/counter/index.web.jsx
@@ -2,15 +2,13 @@ import React from 'react';
 import { Route } from 'react-router-dom';
 
 import Counter from './containers/Counter';
-import ClientCounterFeature, { ClientCounter } from './clientCounter';
-import ReduxCounterFeature, { ReduxCounter } from './reduxCounter';
-import ServerCounterFeature, { ServerCounter } from './serverCounter';
+import clientCounter from './clientCounter';
+import reduxCounter from './reduxCounter';
+import serverCounter from './serverCounter';
 import Feature from '../connector';
 import resources from './locales';
 
-export default new Feature(ClientCounterFeature, ReduxCounterFeature, ServerCounterFeature, {
+export default new Feature(clientCounter, reduxCounter, serverCounter, {
   route: <Route exact path="/" component={Counter} />,
   localization: { ns: 'counter', resources }
 });
-
-export { ClientCounter, ServerCounter, ReduxCounter };
